condensed new component files
